ollama_multi/requirements.txt
langchain-ollama
langchain-community 
langchain-core
